testing database stuff

+++ backend/spring-cashier/src/main/java/com/example/springcashier/PaymentsCommandRepository.java
@@ -2,5 +2,5 @@ package com.example.springcashier;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface PaymentsCommandRepository extends JpaRepository<PaymentsCommand, Long> {
+interface PaymentsCommandRepository extends JpaRepository<PaymentsCommand, Integer> {
 }
